+++ src/components/templates/MainLayout.tsx
@@ -12,7 +12,7 @@ import Logo from '~/static/images/logo.png';
 
 import { DrawerItem, DrawerName } from '~/types/drawer';
 import AppBar from '~/components/organism/appBar';
-import Drawer from '~/components/organism/drawer';
+import Drawer from '~/components/pages/drawer';
 import Profile from '~/components/atom/profile';
 import BottomSheet from '~/components/pages/bottomSheet';
